Refactor out Reporter concern
