+++ app/services/claim_review_async_stats_reporter.rb
@@ -1,8 +1,10 @@
 # frozen_string_literal: true
 
-require "csv"
+# Statistics about AMO decision reviews
 
 class ClaimReviewAsyncStatsReporter
+  include Reporter
+
   attr_reader :stats
 
   def initialize(start_date: Constants::DATES["AMA_ACTIVATION"].to_date, end_date: Time.zone.tomorrow)
@@ -46,11 +48,6 @@ class ClaimReviewAsyncStatsReporter
   end
   # rubocop:enable Metrics/MethodLength
 
-  # public util method
-  def seconds_to_hms(secs)
-    [secs / 3600, secs / 60 % 60, secs % 60].map { |segment| segment.to_s.rjust(2, "0") }.join(":")
-  end
-
   private
 
   attr_reader :start_date, :end_date
@@ -69,8 +66,8 @@ class ClaimReviewAsyncStatsReporter
         established_within_seven_days_percent: percent_established_within_seven_days(
           supplemental_claims_completion_times, supplemental_claims.count
         ),
-        median: median_time(supplemental_claims_completion_times),
-        avg: avg_time(supplemental_claims_completion_times),
+        median: median(supplemental_claims_completion_times),
+        avg: average(supplemental_claims_completion_times),
         max: supplemental_claims_completion_times.max,
         min: supplemental_claims_completion_times.min
       },
@@ -84,8 +81,8 @@ class ClaimReviewAsyncStatsReporter
         established_within_seven_days_percent: percent_established_within_seven_days(
           higher_level_reviews_completion_times, higher_level_reviews.count
         ),
-        median: median_time(higher_level_reviews_completion_times),
-        avg: avg_time(higher_level_reviews_completion_times),
+        median: median(higher_level_reviews_completion_times),
+        avg: average(higher_level_reviews_completion_times),
         max: higher_level_reviews_completion_times.max,
         min: higher_level_reviews_completion_times.min
       },
@@ -99,8 +96,8 @@ class ClaimReviewAsyncStatsReporter
         established_within_seven_days_percent: percent_established_within_seven_days(
           request_issues_updates_completion_times, request_issues_updates.count
         ),
-        median: median_time(request_issues_updates_completion_times),
-        avg: avg_time(request_issues_updates_completion_times),
+        median: median(request_issues_updates_completion_times),
+        avg: average(request_issues_updates_completion_times),
         max: request_issues_updates_completion_times.max,
         min: request_issues_updates_completion_times.min
       }
@@ -155,18 +152,5 @@ class ClaimReviewAsyncStatsReporter
     @request_issues_updates ||= begin
       RequestIssuesUpdate.where("submitted_at >= ? AND submitted_at <= ?", start_date, end_date)
     end
-  end
-
-  def median_time(times)
-    return 0 if times.empty?
-
-    len = times.length
-    (times[(len - 1) / 2] + times[len / 2]) / 2.0
-  end
-
-  def avg_time(times)
-    return 0 if times.empty?
-
-    times.sum.to_f / times.length
   end
 end
