+++ app/services/request_issue_reporter.rb
@@ -3,9 +3,11 @@
 # Reports the number of Request Issues by origin type week-over-week
 
 class RequestIssueReporter
+  include Reporter
+
   attr_reader :stats
 
-  def initialize(start_date: Constants::DATES["AMA_ACTIVATION"].to_date, end_date: Time.zone.today)
+  def initialize(start_date: Constants::DATES["AMA_ACTIVATION"].to_date, end_date: Time.zone.tomorrow)
     @start_date = start_date
     @end_date = end_date
     @stats = build
